+++ src/features/post/components/StepFunnel/style.css.ts
@@ -112,18 +112,6 @@ export const SelectPhotoContainer = css({
   flexWrap: 'wrap',
 });
 
-export const SelectedPhotoBtn = css({
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-  width: '4.38rem',
-  height: '4.38rem',
-  borderRadius: '0.5rem',
-  backgroundColor: 'systemGray5',
-  color: '54',
-  fontSize: '1.875rem',
-});
-
 export const PriceContent = css({
   display: 'flex',
   flex: 1,
