+++ src/features/post/components/UploadPhoto/index.tsx
@@ -1,0 +1,16 @@
+import * as s from './style.css';
+
+interface UploadPhotoProps {
+  onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void;
+}
+
+const UploadPhoto = ({ onChange }: UploadPhotoProps) => {
+  return (
+    <label className={s.SelectedPhotoBtn}>
+      <input type="file" accept="image/*" multiple onChange={onChange} className={s.Input} />
+      <div className="mgc_camera_2_fill" />
+    </label>
+  );
+};
+
+export default UploadPhoto;
