+++ src/features/post/components/UploadPhoto/style.css.ts
@@ -1,0 +1,17 @@
+import { css } from '@styled-system/css';
+
+export const SelectedPhotoBtn = css({
+  display: 'flex',
+  justifyContent: 'center',
+  alignItems: 'center',
+  width: '4.38rem',
+  height: '4.38rem',
+  borderRadius: '0.5rem',
+  backgroundColor: 'systemGray5',
+  color: '54',
+  fontSize: '1.875rem',
+});
+
+export const Input = css({
+  display: 'none',
+});
